secure-storage: rename namespaced_storage.rs to namespaced.rs
Closes: #8429
