+++ secure/storage/src/namespaced.rs
